refactor: raise error if styles are not among the defaults
When doing a GetMap request, the styles validation cannot be performed alone by mapproxy because it's not aware of custom styles defined in each service. At the moment it only supports checking for the `default` layer.

For this reason, changes are added in this PR so if the styles requested are not among the default ones, raise an error so the client of the request can try asking Deegree directly if those styles are valid